# services/blast-analysis/openapi.yaml
openapi: 3.1.0
info:
  title: Drill & Blast AI API
  description: |
    Comprehensive API for drill and blast analysis using 3D point cloud data, 
    DataRobot machine learning, and advanced geometry processing.
    
    ## Features
    - 3D point cloud processing and surface differencing
    - Overbreak/underbreak analysis with volume calculations
    - Hole deviation analysis and burden optimization
    - Fragmentation analysis using computer vision
    - DataRobot AutoML predictions for blast performance
    - Export capabilities (DXF, GeoJSON, CSV, 3D wireframes)
    
    ## Authentication
    All endpoints require JWT Bearer token authentication.
    
    ## Rate Limits
    - 100 requests per hour for analysis submission
    - 1000 requests per hour for status and results
    
  version: 1.0.0
  contact:
    name: GeoVision AI Support
    email: support@geovision.ai
    url: https://geovision.ai/support
  license:
    name: Proprietary
    url: https://geovision.ai/license
servers:
  - url: https://api.geovision.ai/blast-analysis/v1
    description: Production server
  - url: https://staging-api.geovision.ai/blast-analysis/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /blast-analysis/upload:
    post:
      summary: Upload blast data file
      description: |
        Upload point cloud files, CSV data, or DXF designs to S3 storage.
        Returns S3 URI for use in analysis submission.
      operationId: uploadFile
      tags:
        - Data Management
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                blast_id:
                  type: string
                  description: Associated blast ID
                file_type:
                  type: string
                  enum: [preblast, postblast, pattern, loading, mwd, boretrack, design]
                  description: Type of file being uploaded
              required:
                - file
                - blast_id
                - file_type
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '413':
          description: File too large
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UploadResponse:
      type: object
      properties:
        filename:
          type: string
        s3_uri:
          type: string
        size:
          type: integer
        content_type:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information
        code:
          type: string
          description: Error code for programmatic handling
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            details: "Valid JWT token required"
...